Fix a type mismatch warning for function set_memory_wc
v2. Change (long unsigned int) to (unsigned long).

Convert (void *) to (long unsigned int) to match the argument type of
function set_memory_wc.

+++ drivers/gpu/drm/i915/vgt/utility.c
@@ -908,7 +908,7 @@ int setup_gtt(struct pgt_device *pdev)
 		}
 
 		get_page(page);
-		set_memory_wc(page_address(page), 1);
+		set_memory_wc((unsigned long)page_address(page), 1);
 
 		(*pages)[i] = page;
